+++ src/OpenRiaServices.Hosting.AspNetCore/Framework/AspNetCore/Serialization/MessagePack/MessagePackUtility.cs
@@ -28,9 +28,8 @@ namespace OpenRiaServices.Client.DomainClients.MessagePack
 
                 if (derivedTypeShapes != null)
                 {
-                    if (derivedTypeShapes.Name is null)
-                        throw new InvalidOperationException("DerivedTypeShapeAttribute must have a non-null Name property.");
-                    return Encoding.UTF8.GetBytes(derivedTypeShapes.Name);
+                    string name = derivedTypeShapes.Name ?? PolyType.Utilities.ReflectionUtilities.GetDerivedTypeShapeName(type);
+                    return Encoding.UTF8.GetBytes(name);
                 }
                 baseType = baseType.BaseType;
             }